refact: project folder
