--- backend/README.md
@@ -1,7 +1,0 @@
-# mpa2
-Repositório para a implementação de ferramenta de mineração de processos para a disciplina de Seminários em Gerenciamento de Dados e Informação.
-
-# How to run the backend
-```bash
-docker build -t bob . && docker run -d --name bob_container -p 8081:8081 bob
-```
