--- backend/internal/exemplo_besta.py
@@ -1,2 +1,0 @@
-def exemplo(a, b):
-    return a + b
