--- backend/requirements.txt
@@ -1,8 +1,0 @@
-fastapi
-uvicorn
-pydantic
-requests
-pandas
-pm4py
-python-multipart
-graphviz
