+++ frontend/src/Components/Sidebar/Navigator.js
@@ -10,9 +10,10 @@ function Navigator() {
     <div className='nav'>
         <div className='nav-buttons' style={{width: '100%'}}>
         <Button variant="contained">Filtros</Button>
+        <Button variant='contained'>Demandas</Button>
+        <Button variant='contained'>Tarefas</Button>
+
         <ButtonGroup variant="outlined" style={{width: '100%'}} aria-label="outlined button group">
-            <Button>Demandas</Button>
-            <Button>Tarefas</Button>
             <Button>Recursos</Button>
         </ButtonGroup>
         </div>
