+++ frontend/src/data.js
@@ -48,11 +48,3 @@ export const cards = [
         ] 
     }
 ];
-
-export const statistics = {
-    demands: 54,
-    variants: 50,
-    activities: 40, 
-    duration: '120 dias',
-    mean: 30,
-}
