+++ frontend/src/pages/ActivitiesScreen/index.js
@@ -1,0 +1,7 @@
+
+export default function ActivitiesScreen () {
+    return (
+        <>
+        </>
+    );
+}
