+++ frontend/src/pages/DemandsScreen/index.js
@@ -1,0 +1,7 @@
+
+export default function DemandsScreen () {
+    return (
+        <>
+        </>
+    );
+};
