+++ frontend/src/routes.tsx
@@ -1,5 +1,7 @@
 import { BrowserRouter, Routes, Route } from 'react-router-dom'; 
-import MPA2 from './pages/MPA2';
+import FilterScreen from './pages/FilterScreen';
+import DemandsScreen from './pages/DemandsScreen';
+import ActivitiesScreen from './pages/ActivitiesScreen';
 import { Upload } from './pages/Upload';
 import { Links } from './types/enums'
 
@@ -7,8 +9,10 @@ export default function AppRoutes() {
     return (
         <BrowserRouter>
             <Routes>
-                <Route path={Links.Upload} element={<Upload onSuccess={undefined}/>}/>
-                <Route path={Links.MPA2} element={<MPA2 />}/>
+                <Route path={Links.Upload} element={<Upload />} />
+                <Route path={Links.Filter} element={<FilterScreen />} />
+                <Route path={Links.Demands} element={<DemandsScreen />} />
+                <Route path={Links.Activities} element={<ActivitiesScreen />} />
             </Routes>
         </BrowserRouter>
     );
